backend/pyproject.toml
[project]
name = "folo-lite"
version = "0.1.0"
description = "Minimalist RSS Reader using FastAPI, Supabase, and HTMX"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "supabase>=2.10.0",
    "feedparser>=6.0.10",
    "jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
]

[tool.uv]
package = false
